refactor: consolidate OpenCode tests and drop local .gitignore additions
Merge test_config_runner.py, test_runner_selection.py, and
test_opencode_integration.py into a single test_opencode_runner.py,
eliminating duplicate helpers and redundant assertions. Revert .gitignore
to upstream — local planning artifacts don't belong in the PR.

27 tests remain (25 unique, 2 from parametrize); 157 total suite passes.
